pyproject.toml
[project]
name = "ragas-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "google-genai>=1.56.0",
    "jsonref>=1.1.0",
    "ragas>=0.3.9",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
]

ignore = [
    "E501",  # line too long, handled by black
]
